Fixed islamic dates being offset incorrectly
12th of Rabi al-awwal is actually on sunday but I made it saturday which
was wrong. When I removed the offset I added, the weekly view was showing
the incorrect holiday in green because I was checking if the day was an
islamic holiday before adjusting / offsetting and / or adding or removing
any islamic dates. Fixed this by re-checking if the day is an islamic
holiday when I've removed or added any new dates to the islamic date
array.