Prevent displaying tomorrow's islamic holiday
We check if maghrib has ended when deciding to show the islamic holiday
which is bad because if it is after maghrib, I use the next days islamic
holiday. This is weird with days of the week becuase if it is after
maghrib on a Friday, it will show saturday in the holiday / title which
isn't accurate. Changed this so only the islamic date changes but the
day of the week / holiday name remains what it is on that day.