Fix for it breaking when getting next month as int
Decided the next months value to get the next days information using
tomorrow's date and since today is the last day of November and
tomorrow is the 1st day of December, it sets todays month as being 11
but sets the next month as being 13 as it would use tomorrow's date and
add 2 to get the next month.

Added a condition to the variable to stop this from happening.

Pretty sure everything is gonna break at the end of December since it
won't be able to get the next days information once maghrib is reached.
This is because the site only fetches 1 years worth of information, a
tiny flaw in my solution which I didn't have to deal with earlier this
year.....