Fix Array Breaking Site
Originally created an array in which I could add a list of dates to
offset the months with but this seems to break the entire site.
Currently its only possible to add a single date in so will need to
address this so its able to handle a list of dates. Fixed it by just
overwriting a single date